% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ehe.R
\name{available_ehe}
\alias{available_ehe}
\title{Consulta las bases EHE disponibles}
\usage{
available_ehe(encuesta = NULL, localidad = NULL, anio = NULL)
}
\arguments{
\item{encuesta}{Tipo de encuesta.}

\item{localidad}{Localidad o municipio.}

\item{anio}{Año.}
}
\value{
Un tibble con las encuestas disponibles.
}
\description{
Devuelve un catálogo de las encuestas EHE disponibles
por localidad, año y tipo de base.
}
