% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{.catalog}
\alias{.catalog}
\title{Construye el catálogo de la EHE}
\usage{
.catalog()
}
\value{
Un tibble con una fila por recurso.
}
\description{
Obtiene el catálogo desde CKAN y agrega los metadatos derivados.
}
\keyword{internal}
