% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.ckan_request}
\alias{.ckan_request}
\title{Consulta la API de CKAN}
\usage{
.ckan_request(action, ...)
}
\arguments{
\item{action}{Endpoint de la API.}

\item{...}{Parámetros enviados a la API.}
}
\value{
Una lista con el contenido de \code{result}.
}
\description{
Función interna para realizar consultas a la API CKAN.
}
\keyword{internal}
