% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dictionary.R
\name{.parse_dictionary}
\alias{.parse_dictionary}
\title{Procesa un diseno de registro EHE}
\usage{
.parse_dictionary(x)
}
\arguments{
\item{x}{Diseno de registro leido con readxl.}
}
\value{
Un tibble con variables, descripciones, codigos,
etiquetas y tipo de registro.
}
\description{
Funcion interna que transforma el diseno de registro
leido desde Excel en un diccionario de variables y categorias.
}
\keyword{internal}
