% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.parse_resource}
\alias{.parse_resource}
\title{Convierte un recurso de CKAN en un tibble}
\usage{
.parse_resource(resource, dataset)
}
\description{
Convierte un recurso de CKAN en un tibble
}
\keyword{internal}
