% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_demografico.R
\name{ehe_demografico}
\alias{ehe_demografico}
\title{Indicadores sociodemograficos de la EHE}
\usage{
ehe_demografico(base)
}
\arguments{
\item{base}{Base de datos de personas de la EHE.}
}
\value{
Un tibble con los indicadores sociodemograficos.
}
\description{
Calcula indicadores sociodemograficos para la poblacion de
14 anos y mas de la Encuesta de Hogares y Empleo (EHE).
}
\details{
El universo de analisis comprende a las personas de 14 anos
y mas. Los indicadores se calculan utilizando el factor de
expansion \code{ponduni}.

La funcion construye internamente las variables derivadas
\code{etarios} y \code{sexoedad}.
}
\examples{
\dontrun{
ehe_demografico(base)
}

}
