% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_labels.R
\name{ehe_labels}
\alias{ehe_labels}
\title{Etiqueta una base EHE}
\usage{
ehe_labels(
  data,
  dictionary = NULL,
  encuesta = "EHE-M",
  anio = NULL,
  tipo = c("individual", "hogar")
)
}
\arguments{
\item{data}{Base de datos EHE.}

\item{dictionary}{Diccionario obtenido con \code{\link[=ehe_dictionary]{ehe_dictionary()}}.
Si es \code{NULL}, se obtiene automaticamente.}

\item{encuesta}{Encuesta EHE.}

\item{anio}{Ano del diseno de registro.}

\item{tipo}{Tipo de base: \code{"individual"} o \code{"hogar"}.}
}
\value{
La base EHE con etiquetas de variables y valores.
}
\description{
Aplica etiquetas de variables y de valores a una base EHE
utilizando el diseno de registro correspondiente.
}
