% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_metadata.R
\name{ehe_metadata}
\alias{ehe_metadata}
\title{Informacion de una base EHE}
\usage{
ehe_metadata(x)
}
\arguments{
\item{x}{Objeto de clase ehe.}
}
\value{
Un tibble con informacion de la encuesta.
}
\description{
Muestra los metadatos asociados a una base EHE descargada.
}
