% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_tabulate.R
\name{ehe_tabulate}
\alias{ehe_tabulate}
\title{Tabulados ponderados de EHE}
\usage{
ehe_tabulate(base, x, y = NULL, weights = NULL, add.totals = NULL)
}
\arguments{
\item{base}{Base de datos EHE.}

\item{x}{Variable para el tabulado.}

\item{y}{Variable opcional para un tabulado bivariado.}

\item{weights}{Variable de ponderación.}

\item{add.totals}{Indica si se agrega el total.
Puede ser \code{"row"}, \code{"col"}, \code{"both"} o \code{NULL}.}
}
\value{
Un tibble con el tabulado.
}
\description{
Calcula tabulados uni y bivariados utilizando el factor
de expansión correspondiente.
}
