% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.ckan_search}
\alias{.ckan_search}
\title{Obtiene el catálogo de la EHE desde CKAN}
\usage{
.ckan_search()
}
\value{
Lista con los resultados de la consulta.
}
\description{
Obtiene el catálogo de la EHE desde CKAN
}
\keyword{internal}
