% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.download_ckan}
\alias{.download_ckan}
\title{Descarga un archivo desde el portal CKAN}
\usage{
.download_ckan(url, destfile)
}
\arguments{
\item{url}{URL del recurso.}

\item{destfile}{Archivo destino.}
}
\description{
Función interna para descargar recursos.
}
\keyword{internal}
