% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dictionary.R
\name{.parse_dictionary}
\alias{.parse_dictionary}
\title{Procesa un diseño de registro EHE}
\usage{
.parse_dictionary(x)
}
\arguments{
\item{x}{Diseño de registro leído con readxl.}
}
\value{
Un tibble con variables, descripciones, códigos,
etiquetas y tipo de registro.
}
\description{
Función interna que transforma el diseño de registro
leído desde Excel en un diccionario de variables y categorías.
}
\keyword{internal}
