% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_app.R
\name{ehe_app}
\alias{ehe_app}
\title{Shiny application to explore EHE labor indicators}
\usage{
ehe_app()
}
\value{
Does not return an object. Runs a Shiny application.
}
\description{
Opens a Shiny application that allows users to select a municipality
and a year and consult the labor indicators calculated by
\code{\link[=ehe_laboral]{ehe_laboral()}}.
}
\examples{
\dontrun{
ehe_app()
}

}
