% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehe_dictionary.R
\name{ehe_dictionary}
\alias{ehe_dictionary}
\title{Obtiene el diseno de registro EHE}
\usage{
ehe_dictionary(encuesta = "EHE-M", anio, tipo = c("individual", "hogar"))
}
\arguments{
\item{encuesta}{Encuesta EHE.}

\item{anio}{Ano del diseno de registro.}

\item{tipo}{Tipo de base: \code{"hogar"} o \code{"individual"}.}
}
\value{
Un tibble con variables, descripciones,
codigos y etiquetas.
}
\description{
Descarga, lee y procesa el diseno de registro
de una encuesta EHE.
}
