% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ehe.R
\name{get_ehe}
\alias{get_ehe}
\title{Descarga una base EHE}
\usage{
get_ehe(
  encuesta = "EHE-M",
  anio,
  tipo = c("individual", "hogar"),
  localidad = NULL
)
}
\arguments{
\item{encuesta}{Tipo de encuesta. Por defecto "EHE-M".}

\item{anio}{Ano de la encuesta.}

\item{tipo}{Tipo de base: "individual" o "hogar".}

\item{localidad}{Localidad de la encuesta.}
}
\value{
Un tibble con la base descargada.
}
\description{
Descarga una base usuaria de la Encuesta de Hogar y Empleo.
}
